Frag/reassembly: messages bigger than 102 are fragmented
- Downlink traffic is automatically fragmented if > 102
- Uplink traffic ireassembled if fragmented
- Remove incomplete fragment after a given delay if message is not full
 to avoid memory leak (default 0, stay forever)

+++ example/main.c
@@ -25,7 +25,7 @@ char * port_name = "/dev/ttyACM0";
 #define NETWORK_ADDRESS 0xABCDEF
 
 static bool onDataReceived(const uint8_t * bytes,
-                           uint8_t num_bytes,
+                           size_t num_bytes,
                            app_addr_t src_addr,
                            app_addr_t dst_addr,
                            app_qos_e qos,
